Reject an index that is not a whole number
README.md section 4.2 calls `f_(0.5)=1` ill-formed; the interpreter
truncated it to `f_0=1` and said nothing. The conversion also round-tripped
through `size_t`, so a negative index survived only by accident.

+++ include/inkamath/numeric_interface.hpp
@@ -78,7 +78,7 @@ struct numeric_interface_imp<std::complex<T>,false>
 
     static int toInt(const std::complex<T>& a)
     {
-        return static_cast<size_t>(numeric_interface<T>::toInt(a.real()));
+        return numeric_interface<T>::toInt(a.real());
     }
 
     static std::string toString(const std::complex<T>& a)
